requirements.txt
amqp==5.1.1
aniso8601==9.0.1
applicationinsights==0.11.10
atomicwrites==1.4.0
attrs==21.4.0
azure-core==1.24.1
azure-identity==1.10.0
bidict==0.22.0
billiard==3.6.4.0
blinker==1.4
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.29.0
colorama==0.4.5
cryptography==37.0.2
Flask==2.1.2
Flask-Admin==1.6.0
Flask-Authorize==0.2.6
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-RESTful==0.3.9
Flask-SocketIO==5.2.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
google-api-core==2.8.2
google-auth==2.8.0
googleapis-common-protos==1.56.3
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
msal==1.18.0
msal-extensions==1.0.0
opencensus==0.9.0
opencensus-context==0.1.2
opencensus-ext-azure==1.1.4
packaging==21.3
pluggy==1.0.0
portalocker==2.4.0
prompt-toolkit==3.0.29
protobuf==4.21.1
psutil==5.9.1
psycopg2==2.9.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.20.0
python-engineio==4.3.2
python-socketio==5.6.0
pytz==2022.1
requests==2.28.0
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.37
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.1.2
WTForms==3.0.1
